Preu tret de desplaçament, passa a estar a Activitat Algoritme Backtracking fet (Falten fer els metodes xD) Punt i coma a Circuit Canviat consultor preu en float que mhavia deixat a PuntInteres

+++ src/PuntInteres.java
@@ -40,7 +40,7 @@ public abstract class PuntInteres {
     /** @brief Consulta el preu del punt d'interès
 	@pre cert
 	@post retorna el preu */
-    public float preu(){ return preu; }
+    public int preu(){ return preu; }
     
     /** @brief Consulta la posició del punt d'interès
 	@pre cert
